Move color logic to its own file
